service/requirements.txt
requests
sqlalchemy
python-dotenv
python-jose
psycopg2-binary
python-multipart